Move axes setup to coord classes
(cherry picked from commit 7a65e2803a51161c3f4b78a95793b2b04ca67abd)
